requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
openai==1.3.7
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
langchain==0.0.350
langchain-community==0.0.2
langchain-text-splitters==0.0.1
langchain-openai==0.0.5
chromadb==0.4.18
sentence-transformers==2.2.2
huggingface-hub==0.19.4
transformers==4.35.2 